feat : 컨텐츠 상세조회 응답에 Category 정보 추가
