+++ adapters/in-web/src/main/kotlin/com/pokit/content/dto/response/ContentResponse.kt
@@ -1,7 +1,8 @@
 package com.pokit.content.dto.response
 
-import com.fasterxml.jackson.annotation.JsonFormat
+import com.pokit.content.model.CategoryInfo
 import com.pokit.content.model.Content
+import java.time.format.DateTimeFormatter
 
 data class ContentResponse(
     val contentId: Long,
@@ -10,7 +11,17 @@ data class ContentResponse(
     val title: String,
     val memo: String,
     val alertYn: String,
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:")
+    val createdAt: String,
+    val favorites: Boolean = false
+)
+
+data class GetContentResponse(
+    val contentId: Long,
+    val category: CategoryInfo,
+    val data: String,
+    val title: String,
+    val memo: String,
+    val alertYn: String,
     val createdAt: String,
     val favorites: Boolean = false
 )
@@ -25,13 +36,17 @@ fun Content.toResponse() = ContentResponse(
     createdAt = this.createdAt.toString()
 )
 
-fun GetContentResponse.toResponse() = ContentResponse(
-    contentId = this.contentId,
-    categoryId = this.categoryId,
-    data = this.data,
-    title = this.title,
-    memo = this.memo,
-    alertYn = this.alertYn,
-    createdAt = this.createdAt.toString(),
-    favorites = this.favorites
-)
+fun GetContentResult.toResponse(): GetContentResponse {
+    val formatter = DateTimeFormatter.ofPattern("yyyy.MM.dd")
+
+    return GetContentResponse(
+        contentId = this.contentId,
+        category = this.category,
+        data = this.data,
+        title = this.title,
+        memo = this.memo,
+        alertYn = this.alertYn,
+        createdAt = this.createdAt.format(formatter),
+        favorites = this.favorites
+    )
+}
